Crate metered_channel

Metered variant of mpsc channels to be able to extract metrics.

Modules

oneshot

Metered variant of oneshot channels to be able to extract delays caused by delayed responses.

Structs

Meter

A peek into the inner state of a meter.

MeteredReceiver

A receiver tracking the messages consumed by itself.

MeteredSender

The sender component, tracking the number of items sent across it.

Readout

A readout of sizes from the meter. Note that it is possible, due to asynchrony, for received to be slightly higher than sent.

UnboundedMeteredReceiver

A receiver tracking the messages consumed by itself.

UnboundedMeteredSender

The sender component, tracking the number of items sent across it.

Functions

channel

Create a wrapped mpsc::channel pair of MeteredSender and MeteredReceiver.

unbounded

Create a wrapped mpsc::channel pair of MeteredSender and MeteredReceiver.